tmux_manager/pyproject.toml
# filepath: /pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tmux_manager"
version = "100.0.1"
authors = [
    { name = "Tmux Manager Authors", email = "user@example.com" },
]

description = "A comprehensive tmux service manager for reliable session persistence"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Utilities",
]

dependencies = [
    "pyyaml>=5.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",     # For running unit tests
    "black>=23.0",     # For code formatting
    "flake8>=6.0",     # For linting
    "mypy>=1.0",       # For static type checking
    "pre-commit>=3.0", # For pre-commit hooks
]

[project.scripts]
tmux_manager = "tmux_manager.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 140
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = ["pytest.*"]
ignore_missing_imports = true

[tool.flake8]
max-line-length = 140
extend-ignore = "E203"
exclude = [".git", "__pycache__", "build", "dist"]

[tool.autoinit]
src_dir = "src"
ignore_dirs = ["tests", "docs", "examples"]
